implemented MakeCube & refactored method name in mesh class to setGLmode

+++ src/MeshFactory.h
@@ -10,7 +10,7 @@ public:
 
 	Mesh MakeRectangle(float w, float h, float z = 0.0f) const;
 
-	Mesh MakeCube(int l) const;
+	Mesh MakeCube(float l) const;
 
 };
